+++ R/scale_color_flat.R
@@ -6,6 +6,8 @@
 #'
 #' @inheritParams palette_flat
 #' @param discrete Boolean indicating whether color aesthetic is discrete or not.
+#' @param aesthetics A vector of names of the aesthetics that this scale
+#'   should be applied to (e.g., `c('color', 'fill')`).
 #' @param ... Additional arguments passed to `discrete_scale()` when `discrete`
 #'   is `TRUE` or to `scale_color_gradientn()` when `discrete` is `FALSE`.
 #'
@@ -28,13 +30,13 @@
 #'   theme_modern() +
 #'   scale_color_flat_c(palette = "rainbow")
 #' @export
-scale_color_flat <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
+scale_color_flat <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
   pal <- palette_flat(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("flat_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("flat_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -45,14 +47,14 @@ scale_color_flat <- function(palette = "contrast", discrete = TRUE, reverse = FA
 
 #' @rdname scale_color_flat
 #' @export
-scale_color_flat_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
-  scale_color_flat(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_flat_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_flat(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_flat
 #' @export
-scale_color_flat_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, ...) {
-  scale_color_flat(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_flat_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_flat(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_flat
@@ -77,27 +79,27 @@ scale_colour_flat_d <- scale_color_flat_d
 
 #' @rdname scale_color_flat
 #' @export
-scale_fill_flat <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
+scale_fill_flat <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
   pal <- palette_flat(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("flat_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("flat_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
 
 #' @rdname scale_color_flat
 #' @export
-scale_fill_flat_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, ...) {
-  scale_fill_flat(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_flat_d <- function(palette = "contrast", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_flat(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_flat
 #' @export
-scale_fill_flat_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, ...) {
-  scale_fill_flat(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_flat_c <- function(palette = "contrast", discrete = FALSE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_flat(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }