+++ R/scale_color_metro.R
@@ -27,13 +27,13 @@
 #'   theme_modern() +
 #'   scale_color_metro_c(palette = "rainbow")
 #' @export
-scale_color_metro <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
+scale_color_metro <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
   pal <- palette_metro(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("metro_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("metro_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -44,14 +44,14 @@ scale_color_metro <- function(palette = "complement", discrete = TRUE, reverse =
 
 #' @rdname scale_color_metro
 #' @export
-scale_color_metro_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
-  scale_color_metro(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_metro_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_metro(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_metro
 #' @export
-scale_color_metro_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, ...) {
-  scale_color_metro(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_metro_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_metro(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_metro
@@ -76,27 +76,27 @@ scale_colour_metro_d <- scale_color_metro_d
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
+scale_fill_metro <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
   pal <- palette_metro(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("metro_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("metro_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
-  scale_fill_metro(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_metro_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_metro(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, ...) {
-  scale_fill_metro(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_metro_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_metro(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }