+++ R/scale_color_see.R
@@ -28,13 +28,14 @@
 scale_color_see <- function(palette = "contrast",
                             discrete = TRUE,
                             reverse = FALSE,
+                            aesthetics = "color",
                             ...) {
   pal <- palette_see(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("see_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("see_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -48,11 +49,13 @@ scale_color_see <- function(palette = "contrast",
 scale_color_see_d <- function(palette = "contrast",
                               discrete = TRUE,
                               reverse = FALSE,
+                              aesthetics = "color",
                               ...) {
   scale_color_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -62,11 +65,13 @@ scale_color_see_d <- function(palette = "contrast",
 scale_color_see_c <- function(palette = "contrast",
                               discrete = FALSE,
                               reverse = FALSE,
+                              aesthetics = "color",
                               ...) {
   scale_color_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -96,13 +101,14 @@ scale_colour_see_d <- scale_color_see_d
 scale_fill_see <- function(palette = "contrast",
                            discrete = TRUE,
                            reverse = FALSE,
+                           aesthetics = "fill",
                            ...) {
   pal <- palette_see(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("see_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("see_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -112,11 +118,13 @@ scale_fill_see <- function(palette = "contrast",
 scale_fill_see_d <- function(palette = "contrast",
                              discrete = TRUE,
                              reverse = FALSE,
+                             aesthetics = "fill",
                              ...) {
   scale_fill_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }
@@ -126,11 +134,13 @@ scale_fill_see_d <- function(palette = "contrast",
 scale_fill_see_c <- function(palette = "contrast",
                              discrete = FALSE,
                              reverse = FALSE,
+                             aesthetics = "fill",
                              ...) {
   scale_fill_see(
     palette = palette,
     discrete = discrete,
     reverse = reverse,
+    aesthetics = aesthetics,
     ...
   )
 }