+++ R/scale_color_social.R
@@ -26,13 +26,13 @@
 #'   theme_modern() +
 #'   scale_color_social_c(palette = "rainbow")
 #' @export
-scale_color_social <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
+scale_color_social <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
   pal <- palette_social(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("colour", paste0("social_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("social_", palette), palette = pal, ...)
   } else {
-    scale_color_gradientn(colours = pal(256), ...)
+    scale_color_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
@@ -43,14 +43,14 @@ scale_color_social <- function(palette = "complement", discrete = TRUE, reverse 
 
 #' @rdname scale_color_social
 #' @export
-scale_color_social_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
-  scale_color_social(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_social_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_social(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_social
 #' @export
-scale_color_social_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, ...) {
-  scale_color_social(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_color_social_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, aesthetics = "color", ...) {
+  scale_color_social(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_social
@@ -75,27 +75,27 @@ scale_colour_social_d <- scale_color_social_d
 
 #' @rdname scale_color_social
 #' @export
-scale_fill_social <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
+scale_fill_social <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
   pal <- palette_social(palette = palette, reverse = reverse)
 
   if (discrete) {
-    discrete_scale("fill", paste0("social_", palette), palette = pal, ...)
+    discrete_scale(aesthetics = aesthetics, paste0("social_", palette), palette = pal, ...)
   } else {
-    scale_fill_gradientn(colours = pal(256), ...)
+    scale_fill_gradientn(colours = pal(256), aesthetics = aesthetics, ...)
   }
 }
 
 
 #' @rdname scale_color_social
 #' @export
-scale_fill_social_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, ...) {
-  scale_fill_social(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_social_d <- function(palette = "complement", discrete = TRUE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_social(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }
 
 #' @rdname scale_color_social
 #' @export
-scale_fill_social_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, ...) {
-  scale_fill_social(palette = palette, discrete = discrete, reverse = reverse, ...)
+scale_fill_social_c <- function(palette = "complement", discrete = FALSE, reverse = FALSE, aesthetics = "fill", ...) {
+  scale_fill_social(palette = palette, discrete = discrete, reverse = reverse, aesthetics = aesthetics, ...)
 }